FiscalizacaoChocolate.java

+++ FiscalizacaoChocolate.java
@@ -37,7 +37,7 @@ public class FiscalizacaoChocolate {
         System.out.println("Preço do ovo de Páscoa: ");
         double priceOvo = teclado.nextDouble();
         System.out.println("Nota de pureza (de 0 a 10): ");
-        int notaP = teclado.nextInt();
+        double notaP = teclado.nextDouble();
 
         if (priceOvo > 100.00 || notaP < 5) {
             System.out.println("REJEITADO PARA EXPORTAÇÃO: Custo-benefício fora dos padrões");
